Moved test to other file
